2020/10/20201021_Laplacian/requirements.txt
numpy~=1.19.2
opencv-python~=4.4.0.44
matplotlib~=3.3.2
Pillow~=7.2.0